1.01 - New - Refatoração: Mover lógica de negócios do Controller para o Service e implementar injeção de dependências via construtor

+++ src/main/java/matheus/projeto_jdbc/Services/CursosService.java
@@ -3,6 +3,7 @@ package matheus.projeto_jdbc.Services;
 import matheus.projeto_jdbc.Models.Cursos;
 import matheus.projeto_jdbc.Repositories.CursosRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import java.util.List;
 import java.util.Optional;
@@ -10,27 +11,60 @@ import java.util.Optional;
 @Service
 public class CursosService {
 
-    @Autowired
     private CursosRepository cursosRepository;
+
+    public CursosService(CursosRepository cursosRepository) {
+        this.cursosRepository = cursosRepository;
+    }
 
     public List<Cursos> findAll() {
         return cursosRepository.findAll();
     }
 
-    public Optional<Cursos> findById(Long id) {
-        return cursosRepository.findById(id);
+    public ResponseEntity<Cursos> findById(Long id) {
+        Optional<Cursos> cursosFindId = cursosRepository.findById(id);
+        if (cursosFindId.isPresent()) {
+            return ResponseEntity.ok(cursosFindId.get());
+        } else {
+            return ResponseEntity.notFound().build();
+        }
     }
 
-    public boolean save(Cursos cursos) {
-        return cursosRepository.save(cursos) > 0;
+    public ResponseEntity<Cursos> save (Cursos cursos){
+        int resultado = cursosRepository.save(cursos);
+        if (resultado > 0) {
+            return ResponseEntity.ok(cursos);
+        }else {
+            return ResponseEntity.badRequest().build();
+        }
     }
 
-    public boolean update(Cursos cursos) {
-        return cursosRepository.update(cursos) > 0;
+    public Cursos update(Long id, Cursos cursosDetails) {
+        Optional<Cursos> cursosOptional = cursosRepository.findById(id);
+
+        if (cursosOptional.isPresent()) {
+            Cursos cursosEncontrados = cursosOptional.get();
+            cursosEncontrados.setNome(cursosDetails.getNome());
+            cursosEncontrados.setConteudo(cursosDetails.getConteudo());
+
+            boolean updated = cursosRepository.update(cursosEncontrados) > 0;
+            if (updated) {
+                return cursosEncontrados;
+            } else {
+
+                return null;
+            }
+        }else {
+
+            return null;
+        }
     }
 
     public boolean deleteById(Long id) {
-        return cursosRepository.deleteById(id) > 0;
+        if (cursosRepository.findById(id).isPresent()) {
+            return cursosRepository.deleteById(id) > 0;
+        }
+        return false;
     }
 }
